+++ guillotina/tests/cache/test_cache_memory.py
@@ -4,7 +4,11 @@ from guillotina.db.transaction import Transaction
 from guillotina.interfaces import ICacheUtility
 from guillotina.tests import mocks
 from guillotina.tests.utils import create_content
+from unittest import mock
+from unittest.mock import MagicMock
+from unittest.mock import Mock
 
+import asyncio
 import pytest
 
 
@@ -140,3 +144,42 @@ async def test_do_not_cache_large_object(guillotina_main):
     loaded = await txn.get(ob.__uuid__)
     assert id(loaded) != id(ob)
     assert loaded.__uuid__ == ob.__uuid__
+
+
+@pytest.fixture(scope="function")
+def mocked_cache_set(loop):
+    with mock.patch("guillotina.contrib.cache.strategy.BasicCache.set") as mock_set:
+        f = asyncio.Future()
+        f.set_result(None)
+        mock_set.return_value = f
+        yield mock_set
+
+
+@pytest.mark.app_settings(DEFAULT_SETTINGS)
+async def test_fill_cache_doesnt_cache_large_objects(guillotina_main, loop, mocked_cache_set):
+    tm = mocks.MockTransactionManager()
+    txn = Transaction(tm)
+    cache = BasicCache(txn)
+
+    # Mock storing an object that is over the limit
+    obj = Mock()
+    obj.__uuid__ = "foo"
+    obj.__serial__ = "bar"
+    obj.__name__ = "ba"
+    obj.__of__ = "bi"
+
+    pickled = MagicMock()
+    pickled.__len__.return_value = BasicCache.max_cache_record_size + 10
+
+    await cache.store_object(obj, pickled)
+
+    # Call fill_cache and check that caching was skipped
+    await cache.fill_cache()
+
+    cache.set.assert_not_called()
+
+    # Now add smaller object and check that is cached
+    pickled.__len__.return_value = 1
+    await cache.store_object(obj, pickled)
+    await cache.fill_cache()
+    cache.set.assert_called_once()